[vim] Repeat comment marker at each line in rst files
Repeat the '.. ' comment marker on each comment line in reStructuredText
files, like this:

  .. first line
  .. second line

Even though it is not required by reStructuredText syntax, the syntax
highlighting is better when each line starts with '.. '. Also the
vim-commentary plugin works better with the markers, e.g. when typing
'\\ap' in normal mode to (un)comment a paragraph of text.