[emacs] disable vc polling; disable vc status in modeline
empty the list of tried vc backends to avoid trying all of those all
the time when opening a files, or when polling for status changes
because of auto-revert-mode. this is clearly visible in strace output.

side effect is that the vc status (e.g. git branch name) is no longer
shown in the mode line. i don't think i'll miss that.