[bash] avoid warnings in prompt about vte helper script
sourcing vte.sh doesn't really do anything in some cases, e.g. logging
in over ssh. unconditionally trying to use __vte_urlencode() afterwards
will result in errors since that function is not defined.

change the logic to only redefine __vte_osc7() if it was defined in the
first place, which solves the issue.